minor renaming
